requirements.txt
django
django-rest-framework
django-allauth
django-rest-auth